Move enablement out of the FRAMEWORK_AGENT phase handler
Enablement is phase-orthogonal: the Coordinator pumps it on every tick
regardless of ``state.phase``, because the failure it repairs -- a
combination that cannot boot, or boots and misses its accuracy floor --
traps the run in PRELUDE, before any phase that could host the repair has
been entered. A comment on the pump records that housing it inside the
perf pump had made it unreachable for exactly that case. It sat in
``phases/framework.py`` for historical reasons only.

Extract it to ``orchestrator/enablement/``: ``params`` (authoring request
construction), ``lane`` (round admission, in-flight tracking, re-arm),
``build`` (off-loop compiled-build escalation and outcome routing) and
``revalidation`` (the genuine baseline that finalises an eval-origin KEEP).
1757 lines leave ``framework.py``, which drops from 5797 to 4019.

Attribute delegation moves to a new ``CoordinatorCollaborator`` base;
``PhaseHandler`` becomes the subclass that additionally says "bound to one
phase", which is what the ``_on_enter_*`` dispatch keys on. The enablement
collaborators subclass the base directly -- being un-phased is the point.

``_maybe_rearm_authored_lane`` stays: it is a three-line dispatcher whose
body is entirely perf-lane, and its enablement branch is one delegated call.
``_framework_gpu_params`` / ``_framework_authoring_lanes_ttl`` also stay --
four perf call sites each -- so enablement reaches them back through the
Coordinator, the same cross-collaborator hop framework already makes into
``phase_explore`` for ``_warm_specialist_params``.

``writeback`` was reading the stall cap three ways: a module-level import
from ``loop.coordinator`` (its origin), plus two lazy re-imports through
``phases.framework`` with an ``except ImportError: 5`` fallback. Collapse to
the module-level import.
