+++ src/hyperloom/inference_optimizer/tests/test_enablement_coordinator_wiring_unit.py
@@ -305,9 +305,9 @@ def _enqueue_self(**state_kw):
     # Admission on the session wall-clock is exercised in test_coordinator_runtime
     # against a real coordinator; here nothing is ever denied for want of budget.
     fake._time_budget_denial_for_action = lambda _action: None
-    from hyperloom.orchestrator.phases.framework import FrameworkPhase
+    from hyperloom.orchestrator.enablement.lane import EnablementLane
 
-    fake._enablement_in_flight = types.MethodType(FrameworkPhase._enablement_in_flight, fake)
+    fake._enablement_in_flight = types.MethodType(EnablementLane._enablement_in_flight, fake)
     # _enablement_in_flight reads the coordinator's ephemeral pending_proposals.
     fake.state = types.SimpleNamespace(pending_proposals={})
     return fake
