+++ src/hyperloom/orchestrator/phases/framework.py
@@ -6,24 +6,16 @@ specialist dispatch, enablement repair, and Critic-review submission/reauthor.""
 
 from __future__ import annotations
 import asyncio
-import hashlib
 import logging as _logging
 import os
 import subprocess
-import time
-import tempfile
-from collections.abc import Callable
 from datetime import datetime, timezone
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 from hyperloom.common.git_safety import safe_directory_args
 
 from . import machine_state as _phase_state
-from ._enablement_artifacts import snapshot_round, write_setting_script
 from ..bus.message_bus import Message
-from ..actions.executors._accuracy_gate import ENABLEMENT_REVALIDATION_REASON
-from ..actions.executors._grid_server_args import merge_server_args
-from ..loop.coordinator_helpers import _dedupe_extra_server_args
 from ..state.shared_state import inject_stack_base_params, resolve_grading_anchor_tput
 
 if TYPE_CHECKING:
@@ -32,7 +24,6 @@ from ..loop.coordinator import (
     DEFAULT_FRAMEWORK_MAX_CANDIDATES,
     PendingProposal,
     _AUTHORED_LANE_MAX_ATTEMPTS,
-    _ENABLEMENT_MAX_STALL,
     _FRAMEWORK_MIN_PER_REPO_TIMEOUT_SEC,
     _framework_config_levers_from_done,
 )
@@ -45,133 +36,6 @@ log = _logging.getLogger(__name__)
 # one interrupted run — a process restart reclaims an in-flight specialist as
 # failed — must not retire it.
 _LOCAL_EXPLORE_MAX_ATTEMPTS: int = 3
-
-
-def _derive_gpu_arch(gpu_type: str) -> str:
-    """Map a gpu_type label to an explicit GFX arch (never silent fallback)."""
-    _MAP = {
-        "mi355x": "gfx950",
-        "mi300x": "gfx942",
-        "mi308x": "gfx942",
-        "mi300": "gfx942",
-        "mi250x": "gfx90a",
-        "mi250": "gfx90a",
-        "mi210": "gfx90a",
-    }
-    gt = (gpu_type or "").strip().lower()
-    for key, arch in _MAP.items():
-        if key in gt:
-            return arch
-    return ""
-
-
-def _repo_matches_targeted_build_component(repo_url: str, component: str) -> bool:
-    """Return whether a repo name is compatible with a targeted-build recipe.
-
-    This deliberately ignores the origin/owner: specialists may discover and
-    request forks on any host.  It only prevents routing an unrelated repository
-    into a component-specific recipe.  An empty URL is valid because each recipe
-    has its own built-in default repository.
-    """
-    repo = (repo_url or "").strip().rstrip("/")
-    if not repo:
-        return True
-    repo_name = repo.rsplit("/", 1)[-1].removesuffix(".git").lower()
-    hints = {
-        "aiter": ("aiter",),
-        "framework_ext": ("aiter",),
-        "sgl_kernel": ("sglang", "sgl-kernel", "sgl_kernel"),
-        "vllm_source": ("vllm",),
-    }
-    return any(hint in repo_name for hint in hints.get(component, ()))
-
-
-def _maybe_build_runtime_candidate(
-    capability_gap: Any,
-    *,
-    framework: str,
-    model: str,
-    gpu_type: str,
-) -> dict[str, Any] | None:
-    """Build a serialized runtime-candidate stack action, or None.
-
-    Returns None when the gap does not require code acquisition, the run is
-    multi-node (single-node-only guard), or the framework adapter cannot produce
-    an evidence-backed candidate. Fully exception-guarded.
-    """
-    if not getattr(capability_gap, "requires_code_acquisition", False):
-        return None
-    try:
-        from ..actions.executors._multi_node_env import is_multi_node
-
-        if is_multi_node():
-            return None
-        from ..framework.adapters import get_adapter
-
-        adapter = get_adapter(framework)
-        action = adapter.build_stack_action(capability_gap, framework=framework, model=model, gpu_type=gpu_type)
-        if action is None:
-            return None
-        return action.to_state()
-    except Exception:  # noqa: BLE001 — candidate construction is best-effort
-        log.debug("enablement: runtime-candidate construction failed", exc_info=True)
-        return None
-
-
-def _enablement_carrier_params(state: Any) -> dict[str, Any]:
-    """eval-origin trigger context threaded to specialist/integrate/build tasks.
-
-    Empty for boot-origin enablement so the boot path is unchanged.
-    """
-    origin = str(state.enablement.origin or "")
-    if not origin:
-        return {}
-    out: dict[str, Any] = {
-        "enablement_origin": origin,
-        "enablement_accuracy_floor": float(state.enablement.accuracy_floor or 0.0),
-    }
-    cfg = str(state.enablement.probe_config_path or "")
-    if cfg:
-        out["enablement_probe_config_path"] = cfg
-    return out
-
-
-def _maybe_build_localization_candidate(
-    capability_gap: Any,
-    *,
-    framework: str,
-    model: str,
-    repo_url: str,
-    candidate_refs: tuple[str, ...],
-) -> dict[str, Any] | None:
-    """Build a serialized localization stack action, or None.
-
-    Returns None when the gap does not require code acquisition, the run is
-    multi-node, there is no merged-PR candidate ref, or the framework adapter
-    cannot localize. The compiled-closure gate runs later in the executor.
-    """
-    if not getattr(capability_gap, "requires_code_acquisition", False):
-        return None
-    ref = next((r for r in (candidate_refs or ()) if str(r).strip()), "")
-    if not ref or not repo_url:
-        return None
-    try:
-        from ..actions.executors._multi_node_env import is_multi_node
-
-        if is_multi_node():
-            return None
-        from ..framework.adapters import get_adapter
-
-        adapter = get_adapter(framework)
-        action = adapter.build_localization_action(
-            capability_gap, framework=framework, model=model, candidate_ref=ref, repo_url=repo_url
-        )
-        if action is None:
-            return None
-        return action.to_state()
-    except Exception:  # noqa: BLE001 — candidate construction is best-effort
-        log.debug("enablement: localization-candidate construction failed", exc_info=True)
-        return None
 
 
 #: Progress-row status for a candidate the Critic rejected. The gate writes it
@@ -1012,870 +876,6 @@ class FrameworkPhase(PhaseHandler):
                     "framework GPU: gpu_research_lane TTL re-source failed; using base TTL",
                 )
         return lanes, ttl
-
-    def _build_enablement_specialist_params(self, launch_log: str, *, attempt: int = 0) -> dict[str, Any] | None:
-        """Build enablement-specialist params from a captured launch failure.
-
-        Classifies the failure (advisory ``kind`` only — see Q1 hardening),
-        plans bridging discovery, and runs a **best-effort** candidate-PR
-        enumeration (network; fully exception-guarded, degrades to repos-only).
-        The mandate itself is rendered downstream by
-        ``_section_enablement_playbook`` from the structured ``enablement_*``
-        params emitted here, so the prompt text is built once, at the point of
-        use. Returns ``None`` **only** when the launch log is blank (nothing to
-        act on); a non-blank log always yields params, even when it classifies
-        as ``UNKNOWN`` — the LLM specialist repairs from the raw log so a
-        brand-new gap type never wedges the run.
-
-        On a retry (``attempt > 0``) the ranked candidate list is *rotated* so a
-        different bridging PR leads, and the notes flag that prior attempts
-        reverted — steering the sub-agent toward a different bridge.
-
-        Args:
-            launch_log: Captured launch / traceback text.
-            attempt: Zero-based dispatch index; drives candidate rotation and a
-                retry hint in the mandate.
-
-        Returns:
-            dict | None: Specialist task params (tagged ``enablement`` +
-            ``framework_agent_authoring``) or ``None``.
-        """
-        text = (launch_log or "").strip()
-        if not text:
-            return None
-        from hyperloom.agents.framework.enablement import EnablementRequest, classify_failure
-        from hyperloom.agents.framework.enablement_ops import build_search_plan
-        from hyperloom.agents.framework.repo_map import repo_url_for_framework
-
-        state = self.shared_state
-        framework = (getattr(state, "framework", "") or "").strip().lower()
-        model = (getattr(state, "model_name", "") or "").strip()
-        repo_url = repo_url_for_framework(framework)
-
-        # Dispatch a specialist for ANY non-blank launch log, even one that
-        # classifies as ``UNKNOWN``: ``kind`` is advisory (routes bridge-repo
-        # hints and labels the mandate), not a gate.
-        signature = classify_failure(text)
-        req = EnablementRequest(
-            framework=framework,
-            model=model or "(target model)",
-            repo_url=repo_url,
-            launch_log=text,
-            gpu_type=(getattr(state, "gpu_type", "") or "").strip().lower(),
-        )
-        plan = build_search_plan(signature, framework_repo_url=repo_url, model=model)
-        candidate_refs = self._discover_enablement_candidate_refs(req, plan)
-        # Lead with a different candidate each attempt (deterministic left-rotation).
-        if candidate_refs and attempt:
-            n = len(candidate_refs)
-            k = attempt % n
-            candidate_refs = candidate_refs[k:] + candidate_refs[:k]
-        # Persist so _maybe_escalate_to_targeted_build can pick the top candidate.
-        state.enablement.candidate_refs = list(candidate_refs)
-        source_context = self._read_enablement_source_context(signature)
-        # For a weight-init failure, fold the checkpoint's ground-truth per-layer
-        # weight inventory into the mandate so the loop self-corrects each retry.
-        weight_facts = self._derive_checkpoint_weight_facts(text)
-        if weight_facts:
-            source_context = (weight_facts + "\n\n" + source_context) if source_context else weight_facts
-        # Progressing patches from prior rounds, re-applied as a base before this
-        # round's patch (serial-gap stacking); author a fix composing on top.
-        base_patches = [str(p) for p in (state.enablement.kept_patches or [])]
-        # Whole-file artifacts kept by prior rounds, re-installed before the
-        # boot the same way patches are re-applied.
-        base_artifacts = list(state.enablement.kept_artifacts or [])
-        # Only prior rounds' *actually-applied* setup commands (recorded by the
-        # specialist and replayed by integrate_patch) stack as a base. No install
-        # command is ever auto-seeded here: an unpinned upgrade of the shared
-        # serving venv is unsafe (CUDA-wheel clobber of ROCm vLLM/torch,
-        # transformers-major skew) and environment/build acquisition is owned by
-        # the isolated targeted-build path + the specialist's own setup_commands.
-        base_setup = [str(c) for c in (state.enablement.setup_commands or [])]
-        # §1b ENABLEMENT PLAYBOOK renders mandate.task_description via _section_enablement_playbook.
-        # notes carries only per-dispatch dynamic context that §1b cannot provide.
-        notes = ""
-        grounding_drops = list(state.enablement.last_grounding_drop_reason or [])
-        spanned_roots = bool(state.enablement.patches_span_multiple_roots)
-        acc_envs = dict((state.enablement.accepted_config or {}).get("extra_envs") or {})
-        acc_args = str((state.enablement.accepted_config or {}).get("extra_server_args") or "").strip()
-        if base_patches or base_setup or base_artifacts or acc_envs or acc_args:
-            progress_bits = []
-            if base_patches:
-                progress_bits.append(f"{len(base_patches)} prior patch(es): {base_patches}")
-            if base_artifacts:
-                art_targets = [a["target"] for a in base_artifacts[:4]]
-                progress_bits.append(f"{len(base_artifacts)} prior artifact(s) installed: {art_targets}")
-            if base_setup:
-                progress_bits.append(f"{len(base_setup)} prior setup command(s): {base_setup}")
-            if acc_envs or acc_args:
-                cfg_bits = []
-                if acc_envs:
-                    cfg_bits.append(f"envs={acc_envs}")
-                if acc_args:
-                    cfg_bits.append(f"args={acc_args!r}")
-                progress_bits.append("accumulated config from prior advanced rounds: " + "; ".join(cfg_bits))
-            notes = (
-                "STACKED ENABLEMENT (progress so far): the following already "
-                "cleared earlier boot crashes and WILL be re-applied/re-run as a "
-                "base before your changes — do NOT redo them; fix only the CURRENT "
-                "(deeper) failure, composing on top. " + "; ".join(progress_bits)
-            )
-        elif attempt:
-            notes = (
-                f"RETRY (attempt {attempt + 1}): a previous enablement patch for this "
-                f"failure was REVERTED (did not make the combo runnable). Try a DIFFERENT "
-                f"bridging approach / candidate than before."
-            )
-        if grounding_drops:
-            drop_note = (
-                "PATCH GROUNDING FAILURE: the patches the prior round submitted "
-                "were dropped because their target file(s) do not exist in ANY "
-                "allowlisted framework source tree. Targets: "
-                + "; ".join(grounding_drops[:4])
-                + ". Verify the target path with Glob/Read inside the worktree "
-                "before authoring the diff."
-            )
-            notes = (drop_note + "\n\n" + notes).strip() if notes else drop_note
-        if spanned_roots:
-            span_note = (
-                "CROSS-TREE PATCH SET: the prior round's patches targeted more "
-                "than one framework source tree, which integrate_patch cannot "
-                "apply as one set. Submit patches for a single tree per round."
-            )
-            notes = (span_note + "\n\n" + notes).strip() if notes else span_note
-        gap_cid = f"gap.enablement.{signature.kind}"
-        from hyperloom.agents.framework.enablement import CapabilityGap
-
-        capability_gap = CapabilityGap.from_signature(signature)
-
-        # When the gap requires code acquisition (not a resource constraint) and
-        # an adapter can build an evidence-backed candidate, attach a
-        # ``runtime_candidate`` so integrate_patch provisions an attempt-scoped
-        # runtime before booting. Skipped in multi-node mode.
-        runtime_candidate = _maybe_build_runtime_candidate(
-            capability_gap, framework=framework, model=model, gpu_type=req.gpu_type
-        )
-        # When a merged-PR candidate exists, attach a ``localization_candidate``
-        # so integrate_patch localizes the closure into the source tree
-        # (compiled closures defer to the targeted build at apply).
-        localization_candidate = _maybe_build_localization_candidate(
-            capability_gap,
-            framework=framework,
-            model=model,
-            repo_url=repo_url,
-            candidate_refs=tuple(candidate_refs),
-        )
-
-        params_out: dict[str, Any] = {
-            "domain": "enablement_specialist",
-            "gap_canonical_id": gap_cid,
-            "gap_symptom": (f"{framework or '?'} cannot launch {model or 'the target model'}: {signature.kind}"),
-            "gap_layer": "framework",
-            "gap_evidence": {"model": model, "failure_kind": signature.kind},
-            "framework": framework,
-            # Enablement tag routes the integrate gate to runnable_decision.
-            "framework_agent_authoring": True,
-            "enablement": True,
-            "enablement_attempt": attempt,
-            "enablement_failure_kind": signature.kind,
-            "enablement_search_repos": list(plan.repos),
-            # Pre-patch failure signature, replayed by integrate_patch.
-            "enablement_before_signature": signature.to_dict(),
-            # CapabilityGap projection: marks resource_constraint as not actionable.
-            "enablement_capability_gap": capability_gap.to_dict(),
-            "enablement_candidate_refs": list(candidate_refs),
-            # Source lines near the offending site, plus (on a weight-init
-            # failure) the checkpoint's per-layer weight inventory. Rendered
-            # into the mandate by _section_enablement_playbook.
-            "enablement_source_context": source_context,
-            # Progressing patches from prior rounds, stacked as a base.
-            "enablement_base_patches": base_patches,
-            # Whole-file artifact records from prior rounds, re-installed before boot.
-            "enablement_base_artifacts": base_artifacts,
-            # Allowlisted setup commands from prior rounds, replayed before boot.
-            "enablement_setup_commands": base_setup,
-            # Config accumulated by prior advanced rounds. The bench variant
-            # layers this round's proposal on top, so a KEEP is graded on the
-            # whole stack and its effective_config records the whole stack.
-            "base_extra_envs": acc_envs,
-            "base_extra_args": acc_args,
-            "launch_probe": req.launch_probe,
-            "source": "coordinator_internal",
-            "notes": notes,
-            # Whole-machine GPU request. Empty on multi-node / no-GPU hosts.
-            **self._framework_gpu_params(),
-            # eval-origin trigger context (empty for boot-origin enablement).
-            **_enablement_carrier_params(state),
-        }
-        if runtime_candidate is not None:
-            params_out["runtime_candidate"] = runtime_candidate
-        # Re-activate a prior KEEP'd attempt runtime so serial stacking runs on
-        # the same runtime the last round promoted.
-        kept_action = state.enablement.kept_stack_action
-        if isinstance(kept_action, dict) and kept_action and "runtime_candidate" not in params_out:
-            params_out["runtime_candidate"] = kept_action
-        if localization_candidate is not None:
-            params_out["localization_candidate"] = localization_candidate
-        # Inject the last targeted-build failure into the mandate.
-        last_build_failure = state.enablement.last_build_failure or {}
-        if isinstance(last_build_failure, dict) and last_build_failure:
-            fc = str(last_build_failure.get("failure_class") or "")
-            fs = str(last_build_failure.get("failure_summary") or "")
-            if fc or fs:
-                build_note = (
-                    f"PREVIOUS TARGETED-BUILD ATTEMPT: failure_class={fc!r}"
-                    + (f"; {fs}" if fs else "")
-                    + "\nIf the build ran out of time (failure_class='timeout' or "
-                    "'preflight_budget'), request more build_budget_sec or a smaller "
-                    "component scope.  For compile/symbol defects, choose a different "
-                    "ref or narrow the build target."
-                )
-                notes = build_note + "\n\n" + notes
-                params_out["notes"] = notes
-            params_out["enablement_last_build_failure"] = dict(last_build_failure)
-        return params_out
-
-    def _read_enablement_source_context(self, signature: Any, *, window: int = 12) -> str:
-        """Best-effort read a small source window near the offending site.
-
-        Resolves ``signature.offending_file`` against the framework/ROCm source
-        allowlist, then returns ``window`` lines centred on the first occurrence
-        of ``offending_symbol`` (or the file head when the symbol is absent).
-        Fully exception-guarded: any failure returns ``""`` so the mandate
-        degrades to the no-context form (G is grounding, never a hard dependency).
-
-        Delegates to :func:`~..actions.executors._apply_feedback.source_context_for_file`
-        which is the shared file-resolve + window primitive.
-
-        Args:
-            signature: The classified :class:`FailureSignature`.
-            window: Total number of lines to return around the hit.
-
-        Returns:
-            str: A ``file:line`` header + snippet, or ``""``.
-        """
-        offending_file = str(getattr(signature, "offending_file", "") or "").strip()
-        if not offending_file:
-            return ""
-        symbol = str(getattr(signature, "offending_symbol", "") or "").strip()
-        from ..actions.executors._apply_feedback import source_context_for_file
-        from ..framework.paths import resolve_source_file_allowlist
-
-        search_roots = [Path(str(r)) for r in resolve_source_file_allowlist()]
-        return source_context_for_file(
-            offending_file,
-            symbol=symbol,
-            window=window,
-            search_roots=search_roots,
-        )
-
-    def _derive_checkpoint_weight_facts(self, launch_log: str) -> str:
-        """Auto-derive ground-truth checkpoint-weight facts for a weight-init failure.
-
-        On a weight-loading error (strict init / state_dict mismatch), parses the
-        offending ``model...`` parameter names from the launch log and
-        cross-references the model's ``*.index.json`` ``weight_map`` to report,
-        per offending family, which layer indices carry that weight in the
-        checkpoint and which do not. Returns a compact FACTS block appended to the
-        enablement mandate. Fully exception-guarded: any failure returns ``""``.
-
-        Args:
-            launch_log: The captured launch / traceback text.
-
-        Returns:
-            str: A ``CHECKPOINT WEIGHT FACTS`` block, or ``""``.
-        """
-        text = launch_log or ""
-        low = text.lower()
-        try:
-            import glob as _glob
-            import json as _json
-            import re as _re
-            from pathlib import Path as _Path
-
-            # Offending parameter names in the traceback; parsed first so the
-            # trigger is robust to a head-truncated launch log.
-            offending = set(_re.findall(r"['\"]((?:model|language_model|transformer)\.[\w.]+)['\"]", text))
-            weighty = {o for o in offending if o.endswith((".weight", ".bias", "_scale"))}
-            phrase_hit = (
-                "not initialized from checkpoint" in low
-                or "missing key" in low
-                or "unexpected key" in low
-                or "error(s) in loading state_dict" in low
-            )
-            # Fire when the log names weight-shaped offending params even if the
-            # explanatory phrase was truncated off.
-            if not (phrase_hit or weighty):
-                return ""
-            if not offending:
-                return ""
-            model_path = str(getattr(self.shared_state, "model_path", "") or "").strip()
-            if model_path:
-                # ``model_path`` may be an HF repo id; resolve to the local
-                # weights dir so the sharded-index read works for repo-id launches.
-                from hyperloom.inference_optimizer.model_config_utils import (
-                    resolve_local_model_dir,
-                )
-
-                _resolved = resolve_local_model_dir(model_path)
-                if _resolved is not None:
-                    model_path = str(_resolved)
-            if not model_path or not _Path(model_path).is_dir():
-                return ""
-            # Load the checkpoint weight_map (sharded index) or list single-file keys.
-            weight_map: dict[str, Any] = {}
-            idx_files = _glob.glob(f"{model_path}/*.index.json")
-            if idx_files:
-                data = _json.loads(_Path(idx_files[0]).read_text(errors="replace"))
-                weight_map = data.get("weight_map", {}) if isinstance(data, dict) else {}
-            if not weight_map:
-                return ""
-            ckpt_keys = set(weight_map.keys())
-
-            # Group offending names by a layer-index-stripped "family".
-            def _family(name: str) -> str:
-                return _re.sub(r"\.\d+\.", ".{N}.", name)
-
-            def _layer_idx(name: str) -> int | None:
-                m = _re.search(r"\.(\d+)\.", name)
-                return int(m.group(1)) if m else None
-
-            fams: dict[str, dict[str, Any]] = {}
-            for nm in offending:
-                fam = _family(nm)
-                d = fams.setdefault(fam, {"missing_layers": set()})
-                li = _layer_idx(nm)
-                if li is not None:
-                    d["missing_layers"].add(li)
-            lines: list[str] = []
-            for fam in sorted(fams):
-                # Which layer indices for this family exist in the checkpoint?
-                fam_re = _re.compile("^" + _re.escape(fam).replace(r"\{N\}", r"\d+") + "$")
-                present_layers = sorted(
-                    {li for k in ckpt_keys if fam_re.match(k) for li in [_layer_idx(k)] if li is not None}
-                )
-                missing_layers = sorted(fams[fam]["missing_layers"])
-                if present_layers:
-                    lines.append(
-                        f"- '{fam}': PRESENT in checkpoint for layers {present_layers}; "
-                        f"MISSING (instantiated by model, absent from checkpoint) for layers "
-                        f"{missing_layers}. To satisfy the strict init check, the missing layers "
-                        f"must obtain this tensor from a present layer (copy/alias from the nearest "
-                        f"preceding present layer) OR the model must not instantiate it there."
-                    )
-                else:
-                    lines.append(
-                        f"- '{fam}': NOT present in the checkpoint for ANY layer "
-                        f"(missing for layers {missing_layers}). The checkpoint has no source for "
-                        f"this tensor — the model should not require it (guard/skip its "
-                        f"instantiation) rather than copy it."
-                    )
-            if not lines:
-                return ""
-            header = (
-                "CHECKPOINT WEIGHT FACTS (auto-derived from the model's "
-                "safetensors index — GROUND TRUTH, prefer over assumptions). The "
-                "boot failed on weight initialization; for each offending tensor "
-                "family, here is exactly which layers carry it in the checkpoint:"
-            )
-            footer = (
-                "IMPORTANT: verify the exact model class + its load_weights() entry "
-                "point actually used for this architecture (grep the framework "
-                "source for the architecture/model_type) and confirm the parameter-"
-                "dict key naming (with/without a 'model.' prefix) at that scope "
-                "BEFORE writing copy logic — a prior fix silently no-op'd because "
-                "it edited the wrong loader / used mismatched key names, so the copy "
-                "never executed and the SAME weights stayed uninitialized."
-            )
-            return header + "\n" + "\n".join(lines) + "\n" + footer
-        except Exception:  # noqa: BLE001 — auto-facts are best-effort grounding
-            log.debug("enablement: checkpoint weight-facts derivation failed", exc_info=True)
-            return ""
-
-    def _discover_enablement_candidate_refs(self, req: Any, plan: Any) -> tuple[str, ...]:
-        """Best-effort enumerate + rank bridging PRs for an enablement failure.
-
-        Enumerates candidate PRs across every repo in ``plan.repos`` (framework
-        + opted-in ROCm/HIP/aiter bridge repos) via the ``sources`` layer, then
-        ranks each :class:`framework_agent.models.Candidate` with
-        ``score_enablement_title`` (per-Candidate so the ref/html_url is
-        preserved) and returns the top ``req.max_search_candidates`` refs
-        (``html_url`` preferred).
-
-        Network + git; **fully exception-guarded**: any failure degrades to an
-        empty tuple so the mandate falls back to repos-only.
-
-        Args:
-            req: The :class:`framework_agent.enablement.EnablementRequest`.
-            plan: The :class:`framework_agent.enablement_ops.EnablementSearchPlan`.
-
-        Returns:
-            tuple[str, ...]: Ranked candidate refs (best first; possibly empty).
-        """
-        from hyperloom.agents.framework.enablement_ops import score_enablement_title
-        from hyperloom.agents.framework.models import Candidate, ExploreRequest
-        from hyperloom.agents.framework.sources import enumerate_candidates
-
-        max_candidates = int(getattr(req, "max_search_candidates", 5) or 5)
-        # Only search primus_cortex when its URL is configured.
-        primus_url = str(os.environ.get("PRIMUS_CORTEX_PR_API") or "").strip()
-        if primus_url:
-            search_modes = ["primus_cortex", "github"]
-            primus_block: dict[str, Any] = {"primus_cortex": {"base_url": primus_url}}
-        else:
-            search_modes = ["github"]
-            primus_block = {}
-
-        collected: list[Candidate] = []
-        for repo in plan.repos:
-            try:
-                explore_req = ExploreRequest.from_dict(
-                    {
-                        "framework": getattr(req, "framework", "") or "sglang",
-                        "repo_url": repo,
-                        "work_dir": str(
-                            getattr(req, "work_dir", None) or (Path(tempfile.gettempdir()) / "framework-agent")
-                        ),
-                        "baseline": {"throughput": 1.0},
-                        "search_perf_prs": True,
-                        "search_modes": search_modes,
-                        "keywords": list(plan.keywords),
-                        "pr_states": ["all"],
-                        "max_search_candidates": max_candidates,
-                        **primus_block,
-                    }
-                )
-                collected.extend(enumerate_candidates(explore_req))
-            except Exception:  # noqa: BLE001 — discovery is best-effort
-                log.debug(
-                    "enablement: candidate discovery failed for repo=%s",
-                    repo,
-                    exc_info=True,
-                )
-                continue
-
-        if not collected:
-            return ()
-        ranked = sorted(
-            collected,
-            key=lambda c: score_enablement_title(getattr(c, "title", "") or "", plan),
-            reverse=True,
-        )
-        refs: list[str] = []
-        seen: set[str] = set()
-        for cand in ranked:
-            ref = str(getattr(cand, "html_url", "") or getattr(cand, "ref", "") or "").strip()
-            if ref and ref not in seen:
-                seen.add(ref)
-                refs.append(ref)
-            if len(refs) >= max_candidates:
-                break
-        return tuple(refs)
-
-    async def _maybe_enqueue_enablement_specialist(self) -> str:
-        """Dispatch an enablement_specialist when a baseline cannot launch or its
-        accuracy eval fails.
-
-        Retries until the combo runs correctly or the run wall-clock deadline
-        passes (no attempt-count cap). Guards:
-
-        * ``enablement_mode`` — the lane matching the trigger origin must be
-          admitted by ``--enablement``; otherwise no round is ever opened.
-        * ``enablement_succeeded`` — terminal: a prior attempt was KEPT.
-        * ``enablement_validation_pending`` — an eval-origin KEEP is awaiting
-          genuine-baseline revalidation; authoring is paused until it resolves.
-        * ``inflight_task_id`` — a round is still resolving (specialist task or
-          the integrate that consumes its deliverable); derived, never stored.
-        * run deadline passed — stop dispatching new work near the close.
-
-        A non-blank log is always dispatched (the specialist repairs from the raw
-        log even when it classifies to ``UNKNOWN``); only a blank log is a no-op,
-        recorded once as ``needs_human_review``. No-op on multi-node.
-
-        Returns:
-            str: The dispatched specialist ``task_id`` (empty when skipped).
-        """
-        from ..actions.executors._accuracy_gate import eval_enablement_allowed, launch_enablement_allowed
-
-        state = self.shared_state
-        origin = str(state.enablement.origin or "")
-        admitted = eval_enablement_allowed(state) if origin == "eval" else launch_enablement_allowed(state)
-        if not admitted:
-            return ""
-        if bool(state.enablement.succeeded):
-            return ""
-        if bool(state.enablement.validation_pending):
-            # A KEEP'd eval-origin patch is awaiting genuine-baseline revalidation;
-            # do not dispatch another authoring round until that resolves.
-            return ""
-        if state.enablement.inflight_task_id:
-            if await self._enablement_in_flight():
-                return ""
-            # Round ended without calling _maybe_rearm_enablement — count as stall.
-            self._maybe_rearm_enablement(
-                {"enablement": True, "status": "reverted", "reason": "round_finished_without_rearm"}
-            )
-            if state.stop_reason:
-                return ""
-        if float(getattr(state, "baseline_tput", 0.0) or 0.0) > 0:
-            return ""
-        if int(getattr(state, "baseline_failure_streak", 0) or 0) < 1:
-            return ""
-        # Stop opening new enablement attempts once the run deadline has passed.
-        deadline = getattr(self, "_run_deadline", None)
-        if deadline is not None and time.monotonic() >= float(deadline):
-            return ""
-        launch_log = str(state.enablement.launch_log or "")
-        attempt = int(state.enablement.attempts or 0)
-        params = self._build_enablement_specialist_params(launch_log, attempt=attempt)
-        if params is None:
-            # A non-blank UNKNOWN log is recorded for human review, once per log.
-            await self._maybe_record_enablement_human_review(launch_log)
-            return ""
-        # Enqueue any build the *previous* round's specialist explicitly requested
-        # (``needs_targeted_build`` in its specialist_done), then auto-escalate to
-        # a targeted build when the residual gap is a compiled miss or a vLLM
-        # arch/weight deep-failure that source patches keep hitting. Both enqueues
-        # are no-ops when a matching build is already queued/running (idempotent by
-        # novelty key).
-        try:
-            await self._maybe_enqueue_specialist_requested_build()
-        except Exception:  # noqa: BLE001 — best-effort; never wedge dispatch
-            log.debug("enablement: specialist-requested build raised", exc_info=True)
-        try:
-            await self._maybe_escalate_to_targeted_build(launch_log, attempt=attempt)
-        except Exception:  # noqa: BLE001 — escalation is best-effort; never wedge dispatch
-            log.debug("enablement: targeted-build escalation raised", exc_info=True)
-        from ..actions.executors._multi_node_env import is_multi_node
-
-        if is_multi_node():
-            return ""
-        try:
-            await self._warm_specialist_params(params)
-        except Exception:  # noqa: BLE001 — best-effort warmup
-            log.debug("enablement: warm specialist params failed", exc_info=True)
-        idem = f"enablement_authoring:{params.get('enablement_failure_kind', '')}:{attempt}"
-        # This internal dispatch bypasses intent_router (adds gpu_research_lane + budget TTL).
-        lanes, ttl = self._framework_authoring_lanes_ttl(params, base_ttl_sec=3600)
-        spec_task, _existing = await self.tasks.create_or_return_existing(
-            kind="specialist",
-            params=params,
-            idempotency_key=idem,
-            requires_lanes=lanes,
-            side_effects=["writes_results", "writes_patches"],
-            lease_ttl_sec=ttl,
-        )
-        spec_tid = str(getattr(spec_task, "task_id", "") or "")
-        state.enablement.attempts = attempt + 1
-        state.enablement.inflight_task_id = spec_tid
-        try:
-            state.save(self.session_dir)
-        except Exception:  # noqa: BLE001 — defensive
-            log.debug("enablement: save after dispatch failed", exc_info=True)
-        log.info(
-            "ENABLEMENT: dispatched authoring specialist kind=%s attempt=%d task=%s",
-            params.get("enablement_failure_kind"),
-            attempt + 1,
-            spec_tid,
-        )
-        return spec_tid
-
-    async def _enablement_in_flight(self) -> bool:
-        """True while the current enablement round has not settled.
-
-        A round spans the authoring specialist AND the ``integrate_patch`` that
-        consumes its deliverable; the specialist goes terminal a tick before the
-        Critic sees the integrate proposal, so its task state alone is not enough.
-        """
-        from ..state.task_registry import TaskNotFound
-
-        tid = self.shared_state.enablement.inflight_task_id
-        if not tid:
-            return False
-        try:
-            spec = await self.tasks.get(tid)
-        except TaskNotFound:
-            spec = None
-        if spec is not None and spec.state in ("queued", "running"):
-            return True
-        # Undecided proposal keeps the round open; once ruled, approve lands the
-        # task matched below and reject rearms directly, so it cannot defer forever.
-        for p in self.state.pending_proposals.values():
-            if p.action_name != "integrate_patch" or p.decided:
-                continue
-            if (p.payload.get("params") or {}).get("specialist_task_id") == tid:
-                return True
-        for t in (await self.tasks.queued()) + (await self.tasks.running()):
-            if t.kind == "integrate_patch" and t.params.get("specialist_task_id") == tid:
-                return True
-        return False
-
-    async def _maybe_record_enablement_human_review(self, launch_log: str) -> None:
-        """Record a one-shot ``needs_human_review`` for an UNKNOWN launch failure.
-
-        The enablement path only dispatches authoring for *actionable* failure
-        signatures; a non-blank log that classifies to ``UNKNOWN`` used to be
-        silently dropped. Instead, emit a single observation
-        (deduped per distinct log via a stored hash) carrying the classified
-        signature (``raw_excerpt`` + ``offending_file``) so an operator can pick
-        it up. No sub-agent is dispatched.
-
-        Args:
-            launch_log: The captured launch / traceback text.
-        """
-        text = (launch_log or "").strip()
-        if not text:
-            return
-
-        from hyperloom.agents.framework.enablement import classify_failure
-
-        signature = classify_failure(text)
-        if signature.is_actionable:
-            return
-        digest = hashlib.sha1(text.encode("utf-8", errors="replace"), usedforsecurity=False).hexdigest()
-        state = self.shared_state
-        seen = state.enablement.human_review_logged
-        if not isinstance(seen, list):
-            seen = []
-            state.enablement.human_review_logged = seen
-        if digest in seen:
-            return
-        seen.append(digest)
-        framework = (getattr(state, "framework", "") or "").strip().lower()
-        model = (getattr(state, "model_name", "") or "").strip()
-        try:
-            await self._record_observation(
-                "coordinator",
-                "observation",
-                {
-                    "kind": "enablement_needs_human_review",
-                    "applicability": "needs_human_review",
-                    "framework": framework,
-                    "model": model,
-                    "failure_kind": signature.kind,
-                    "signature": signature.to_dict(),
-                    "reason": (
-                        "baseline launch failure did not match any actionable enablement signature; needs human triage"
-                    ),
-                },
-            )
-        except Exception:  # noqa: BLE001 — observability is best-effort
-            log.debug("enablement: human-review record failed", exc_info=True)
-        try:
-            state.save(self.session_dir)
-        except Exception:  # noqa: BLE001 — defensive
-            log.debug("enablement: save after human-review failed", exc_info=True)
-        log.info(
-            "ENABLEMENT: recorded needs_human_review for UNKNOWN failure kind=%s",
-            signature.kind,
-        )
-
-    def _maybe_rearm_enablement(self, res: dict[str, Any] | None) -> None:
-        """Re-arm, advance, or terminate the enablement retry loop.
-
-        Called on every ``integrate_patch`` completion. For an enablement patch
-        there are three outcomes:
-
-        * ``kept`` — the combo is now fully runnable: terminal success
-          (``enablement_succeeded=True``).
-        * ``advanced`` — the patch cleared the prior crash and the boot now
-          stops at a *new, deeper* gap (serial enablement). **Stack** the patch
-          (append to ``enablement_kept_patches``), replace
-          ``enablement_launch_log`` with the new failure so the next round
-          classifies and targets gap #(n+1), reset the stall streak, and clear
-          the in-flight guard to dispatch the next round.
-        * anything else (``reverted`` / apply / bench failure) — no progress:
-          bump ``enablement_stall_streak``; once it reaches
-          :data:`_ENABLEMENT_MAX_STALL`, stop the run with
-          ``stop_reason='enablement_stalled'`` instead of looping on the same
-          gap; otherwise clear the guard so the next round retries a different
-          approach.
-
-        Args:
-            res: The integrate_patch result dict (may be ``None`` / non-dict).
-        """
-        if not isinstance(res, dict) or not res.get("enablement"):
-            return
-        state = self.shared_state
-        status = str(res.get("status") or "")
-        stop_set = ""
-        # Capture the finished round's specialist task id so the async dispatch
-        # chokepoint can read its specialist_done.json for a needs_targeted_build
-        # request (a build enqueue needs await; rearm is sync). Only overwrite on
-        # a real specialist round (targeted_build rearm rows carry no such id).
-        _spec_tid = str(res.get("specialist_task_id") or "").strip()
-        if _spec_tid:
-            state.enablement.last_specialist_task_id = _spec_tid
-
-        def _stack_setup_commands() -> None:
-            """Append this round's applied setup commands to the durable stack."""
-            cur = list(state.enablement.setup_commands or [])
-            for c in res.get("setup_commands_applied") or []:
-                sc = str(c)
-                if sc and sc not in cur:
-                    cur.append(sc)
-            state.enablement.setup_commands = cur
-
-        def _push_kept_round(patches_this_round: list[str]) -> None:
-            """Append this round to kept_rounds and re-derive the flat projections.
-
-            Artifacts dedupe last-wins per target so a later round supersedes an
-            earlier fix to the same file.
-            """
-            rounds = list(state.enablement.kept_rounds or [])
-            rounds.append(
-                {
-                    "patches": list(patches_this_round),
-                    "artifacts": [dict(a) for a in (res.get("artifacts_applied") or []) if isinstance(a, dict)],
-                }
-            )
-            state.enablement.kept_rounds = rounds
-
-            flat_patches: list[str] = []
-            artifact_by_target: dict[str, dict] = {}
-            for rnd in rounds:
-                for p in rnd.get("patches") or []:
-                    if p not in flat_patches:
-                        flat_patches.append(p)
-                for art in rnd.get("artifacts") or []:
-                    target = str(art.get("target") or "")
-                    if target:
-                        artifact_by_target[target] = art
-            state.enablement.kept_patches = flat_patches
-            state.enablement.kept_artifacts = list(artifact_by_target.values())
-
-        def _reset_baseline_failure_backstop() -> None:
-            """Clear the baseline-failure counters on enablement forward progress.
-
-            A serial enablement makes the baseline re-fail on purpose (each round
-            clears gap #n and the next boot stops at a deeper gap), so those
-            crashes are progress, not a stuck baseline. Reset the backstop
-            counters so ``enablement_stalled`` is the sole enablement-phase
-            fast-fail.
-            """
-            state.baseline_failure_streak = 0
-            state.baseline_arg_error_streak = 0
-            state.baseline_total_failures = 0
-
-        def _stack_kept_runtime() -> None:
-            """Persist the KEEP'd attempt runtime + localization manifest so they
-            survive rearm."""
-            action = res.get("enablement_kept_stack_action")
-            if isinstance(action, dict) and action:
-                state.enablement.kept_stack_action = action
-            runtime = res.get("enablement_active_runtime")
-            if isinstance(runtime, dict) and runtime:
-                state.enablement.active_runtime = runtime
-                # Retain the attempt-runtime record (cap at 5 newest).
-                records = list(state.enablement.attempt_runtimes or [])
-                records.append(runtime)
-                state.enablement.attempt_runtimes = records[-5:]
-            # Record the localized closure manifest so it is not re-fetched on
-            # the next round.
-            manifest = res.get("enablement_localization_manifest")
-            if isinstance(manifest, dict) and manifest:
-                existing = list(state.enablement.localization_manifest or [])
-                existing.append(manifest)
-                state.enablement.localization_manifest = existing
-
-        if status == "kept":
-            _reset_baseline_failure_backstop()
-            _stack_setup_commands()
-            _stack_kept_runtime()
-            _push_kept_round([str(p) for p in (res.get("patches_applied") or []) if str(p)])
-            accepted_cfg = str(res.get("enablement_accepted_config_path") or "").strip()
-            if accepted_cfg:
-                state.enablement.accepted_config_path = accepted_cfg
-            effective = res.get("enablement_effective_config")
-            if isinstance(effective, dict) and effective:
-                # Replaced, not merged: what the KEEP bench launched already
-                # supersedes every advanced round that fed into it.
-                state.enablement.accepted_config = dict(effective)
-            if str(state.enablement.origin or "") == "eval":
-                # eval-origin: the patch boots and re-passed accuracy in the gate,
-                # but tput/accuracy only become official once a GENUINE baseline
-                # promotes. Hold succeeded; open the revalidation window. Keep the
-                # stall streak so repeated KEEP->revalidation-fail cycles still
-                # reach the stall cap.
-                state.enablement.validation_pending = True
-                # Increment generation so the new window gets a fresh idempotency
-                # key and cannot reuse a prior terminal TaskRegistry row.
-                state.enablement.revalidation_generation = int(state.enablement.revalidation_generation or 0) + 1
-                state.enablement.revalidation_task_id = ""
-            else:
-                state.enablement.succeeded = True
-                state.enablement.stall_streak = 0
-        elif status == "advanced" or bool(res.get("advanced")):
-            # Forward progress on a serial enablement: stack the progressing
-            # patches + setup commands and pivot to the newly-revealed gap.
-            _push_kept_round([str(p) for p in (res.get("patches_applied") or []) if str(p)])
-            _stack_setup_commands()
-            _stack_kept_runtime()
-            # Accumulated so a later kept round replays every advance, not just patches.
-            adv_envs = res.get("extra_envs_applied") or {}
-            adv_args = str(res.get("extra_server_args_applied") or "").strip()
-            if adv_envs or adv_args:
-                cfg = dict(state.enablement.accepted_config or {})
-                merged = dict(cfg.get("extra_envs") or {})
-                merged.update({str(k): str(v) for k, v in adv_envs.items()})
-                cfg["extra_envs"] = merged
-                # Folded by flag keeping the last value, so this round overrides an earlier one.
-                cfg["extra_server_args"] = _dedupe_extra_server_args(
-                    merge_server_args(str(cfg.get("extra_server_args") or ""), adv_args)
-                )
-                cfg.setdefault("args_mode", "append")
-                state.enablement.accepted_config = cfg
-            new_log = str(res.get("enablement_launch_log") or "").strip()
-            if new_log:
-                state.enablement.launch_log = new_log
-            state.enablement.stall_streak = 0
-            _reset_baseline_failure_backstop()
-        else:
-            # No progress: count toward the stall cap.
-            state.enablement.stall_streak = int(state.enablement.stall_streak or 0) + 1
-            if state.enablement.stall_streak >= _ENABLEMENT_MAX_STALL and not state.stop_reason:
-                state.set_stop_reason("enablement_stalled")
-                stop_set = "enablement_stalled"
-        # Set on every round so neither outlives the round it describes.
-        state.enablement.last_grounding_drop_reason = [
-            str(d) for d in (res.get("patches_dropped_by_grounding") or [])[:8]
-        ]
-        state.enablement.patches_span_multiple_roots = bool(res.get("patches_span_multiple_roots"))
-        # Phase-synthesised rounds carry no framework_root; keep the last real one.
-        res_fw_root = str(res.get("framework_root") or "").strip()
-        if res_fw_root:
-            state.enablement.framework_root = res_fw_root
-        try:
-            snapshot_round(self.session_dir, res)
-            if status in ("kept", "advanced"):
-                write_setting_script(
-                    self.session_dir,
-                    state.enablement,
-                    framework=str(state.framework or os.environ.get("FRAMEWORK") or "sglang"),
-                    model=os.environ.get("MODEL_PATH") or state.model_path or state.reference_model,
-                    tp=int(state.tp or 0) or None,
-                    max_model_len=int(state.max_model_len or 0) or None,
-                    gpu_type=str(state.gpu_type or os.environ.get("GPU_TYPE") or "") or None,
-                )
-        except Exception:  # noqa: BLE001 — archiving must not break the rearm
-            log.warning("enablement: artifact write failed", exc_info=True)
-        # A rearm always ends the round.
-        state.enablement.inflight_task_id = ""
-        try:
-            state.save(self.session_dir)
-        except Exception:  # noqa: BLE001 — defensive
-            log.debug("enablement: save after rearm failed", exc_info=True)
-        log.info(
-            "ENABLEMENT: rearm from integrate status=%s succeeded=%s advanced=%s "
-            "stacked=%d stall_streak=%d next_attempt=%d%s",
-            status,
-            bool(state.enablement.succeeded),
-            status == "advanced" or bool(res.get("advanced")),
-            len(state.enablement.kept_patches or []),
-            int(state.enablement.stall_streak or 0),
-            int(state.enablement.attempts or 0),
-            f" stop_reason={stop_set}" if stop_set else "",
-        )
 
     def _maybe_rearm_authored_lane(self, res: dict[str, Any] | None) -> None:
         """Unified rearm dispatcher for all authored lanes.
@@ -4013,783 +3013,6 @@ class FrameworkPhase(PhaseHandler):
             await self._pump_framework_agent_phase()
         except Exception:  # noqa: BLE001 — defensive
             log.exception("FRAMEWORK pump (%s) failed", caller)
-
-    async def _maybe_escalate_to_targeted_build(
-        self,
-        launch_log: str,
-        *,
-        attempt: int = 0,
-    ) -> None:
-        """Enqueue a targeted build row when the residual gap is a compiled miss.
-
-        No-op when a build is already queued or running (idempotent by novelty
-        key), when the env var ``HYPERLOOM_ENABLEMENT_DISABLE_TARGETED_BUILD=1``
-        is set, on multi-node, or when the gap is not a compiled miss.
-        """
-        import os as _os
-
-        if _os.environ.get("HYPERLOOM_ENABLEMENT_DISABLE_TARGETED_BUILD", "").strip() == "1":
-            return
-        try:
-            from ..actions.executors._multi_node_env import is_multi_node
-
-            if is_multi_node():
-                return
-        except Exception:  # noqa: BLE001
-            return
-        try:
-            from hyperloom.agents.framework.enablement import (
-                MISSING_MODEL_ARCH,
-                MISSING_WEIGHT,
-                NOT_IMPLEMENTED,
-                classify_failure,
-                is_targeted_build_candidate,
-            )
-            from ..framework.build_actions import TargetedBuildAction
-
-            signature = classify_failure(launch_log)
-
-            state = self.shared_state
-            framework = (getattr(state, "framework", "") or "").strip().lower()
-            gpu_type = (getattr(state, "gpu_type", "") or "").strip().lower()
-
-            # Two escalation triggers:
-            #  1. An inherently *compiled* gap (build / rocm_hip / native-dtype /
-            #     hip-kernel) — the original Rung-5 path.
-            #  2. A vLLM **arch/weight deep-failure that source patches keep
-            #     hitting**: the enablement specialist authored ≥1 source patch
-            #     (attempt >= 1) yet the boot still stops at an arch/weight/
-            #     not-implemented wall. For a genuinely new architecture, aliasing
-            #     to an existing model class in the *installed* vLLM cannot model
-            #     the new op set; the correct acquisition is a from-source vLLM
-            #     build of a version that natively implements the arch. Only fires
-            #     on vLLM (the from-source recipe target) and never on the first
-            #     attempt (give the cheap source-patch path a chance first).
-            is_compiled_gap = is_targeted_build_candidate(signature, launch_log)
-            arch_stall = (
-                signature.kind in (MISSING_MODEL_ARCH, MISSING_WEIGHT, NOT_IMPLEMENTED)
-                and framework == "vllm"
-                and int(attempt or 0) >= 1
-            )
-            if not is_compiled_gap and not arch_stall:
-                return
-
-            # Derive novelty fields from the current failure + session context.
-            # Ref and repo_url come from the existing stack action when present;
-            # otherwise the top discovery candidate for the component is used;
-            # empty ref falls back to tag-descending autoselect.
-            existing_stack = state.enablement.kept_stack_action or {}
-            repo_url = str(existing_stack.get("repo_url") or "").strip()
-            ref = str(existing_stack.get("ref") or "").strip()
-
-            # Pick the component from the failure evidence:
-            # sgl_kernel when the offending symbol/log names sgl-kernel;
-            # vllm_source when vLLM's own C extension is implicated;
-            # aiter (default) for all other compiled-miss gaps.
-            sym_lower = (signature.offending_symbol or "").lower()
-            log_lower = launch_log.lower()
-            if arch_stall and not is_compiled_gap:
-                # Arch/weight deep-failure on vLLM after source patches: the fix
-                # is a from-source vLLM build that natively implements the arch,
-                # not an aiter/sgl-kernel op build.
-                component = "vllm_source"
-            elif "sgl_kernel" in sym_lower or "sgl-kernel" in sym_lower or "sgl_kernel" in log_lower:
-                component = "sgl_kernel"
-            elif framework == "vllm" and (
-                "vllm/_c" in log_lower
-                or "vllm.extension" in log_lower
-                or "_c.so" in log_lower
-                or "vllm._c" in sym_lower
-            ):
-                component = "vllm_source"
-            else:
-                component = "aiter"
-
-            source_pr_url = ""
-            # When no operator-pinned/kept ref, try the top discovery candidate.
-            if not ref:
-                from ..framework.build_actions import resolve_build_ref
-
-                _component_hints = {
-                    "aiter": ("aiter",),
-                    "sgl_kernel": ("sglang", "sgl-kernel", "sgl_kernel"),
-                    "vllm_source": ("vllm",),
-                    "framework_ext": (),
-                }
-                hints = _component_hints.get(component, ())
-                default_repo = repo_url or ""
-                for _cand in list(state.enablement.candidate_refs or []):
-                    _repo, _ref, _pr_url = resolve_build_ref(str(_cand), default_repo)
-                    if not _ref:
-                        continue
-                    if hints:
-                        _repo_lower = (_repo or _cand).lower()
-                        if not any(h in _repo_lower for h in hints):
-                            continue
-                    repo_url = _repo or repo_url
-                    ref = _ref
-                    source_pr_url = _pr_url
-                    break
-
-            reason = (
-                f"Rung-5 arch-stall auto-escalation: {signature.kind} persists on vLLM "
-                f"after {int(attempt or 0)} source-patch attempt(s) — build vLLM from source"
-                if (arch_stall and not is_compiled_gap)
-                else f"Rung-5 auto-escalation from {signature.kind}"
-            )
-            if not _repo_matches_targeted_build_component(repo_url, component):
-                log.warning(
-                    "ENABLEMENT: targeted_build repo/component mismatch component=%s repo_url=%s",
-                    component,
-                    repo_url,
-                )
-                return
-            action = TargetedBuildAction(
-                gap_id=f"gap.enablement.{signature.kind}",
-                framework=framework or "vllm",
-                component=component,
-                capability=str(signature.offending_symbol or signature.kind or ""),
-                reason=reason,
-                repo_url=repo_url,
-                ref=ref,
-                gpu_arch=_derive_gpu_arch(gpu_type),
-                build_budget_sec=0,
-                source_pr_url=source_pr_url,
-            )
-            task_id = await self.enqueue_targeted_build(action)
-            if task_id:
-                log.info(
-                    "ENABLEMENT: enqueued targeted_build kind=%s component=%s arch_stall=%s gpu_arch=%s task=%s",
-                    signature.kind,
-                    component,
-                    bool(arch_stall and not is_compiled_gap),
-                    action.gpu_arch,
-                    task_id,
-                )
-        except Exception:  # noqa: BLE001 — escalation is best-effort; never wedge dispatch
-            log.debug("enablement: targeted-build escalation failed", exc_info=True)
-
-    async def _maybe_enqueue_specialist_requested_build(
-        self,
-        *,
-        task_id: str = "",
-        payload: dict[str, Any] | None = None,
-    ) -> None:
-        """Enqueue a targeted build the enablement specialist explicitly requested.
-
-        The enablement specialist may emit a ``needs_targeted_build`` object in its
-        ``specialist_done.json`` (see ``ENABLEMENT_BUILD_REQUEST_GUIDANCE``) when a
-        compiled component / from-source framework build is required that a source
-        patch against the installed tree cannot deliver. This reads that request
-        from the just-finished round's workdir (task id captured at rearm into
-        ``enablement_last_specialist_task_id``) and enqueues it on the isolated,
-        ROCm-safe build lane. The field is cleared once read so the request is
-        consumed at most once; ``enqueue_targeted_build`` is additionally
-        idempotent by build-novelty key. Best-effort — never wedges dispatch.
-        """
-        import os as _os
-
-        state = self.shared_state
-        marker_task_id = str(state.enablement.last_specialist_task_id or "").strip()
-        task_id = str(task_id or marker_task_id).strip()
-        if not task_id:
-            return
-
-        def _consume_marker() -> None:
-            if marker_task_id == task_id:
-                state.enablement.last_specialist_task_id = ""
-
-        try:
-            if _os.environ.get("HYPERLOOM_ENABLEMENT_DISABLE_TARGETED_BUILD", "").strip() == "1":
-                _consume_marker()
-                return
-            from ..actions.executors._multi_node_env import is_multi_node
-
-            if is_multi_node():
-                _consume_marker()
-                return
-            if payload is None:
-                done_path = self.session_dir / "runs" / "specialist" / task_id / "specialist_done.json"
-                if not done_path.is_file():
-                    return
-                import json as _json
-
-                try:
-                    payload = _json.loads(done_path.read_text())
-                except Exception:  # noqa: BLE001 — malformed/partial done is non-fatal
-                    return
-            req = payload.get("needs_targeted_build") if isinstance(payload, dict) else None
-            if not isinstance(req, dict) or not req:
-                _consume_marker()
-                return
-
-            from ..framework.build_actions import (
-                _COMPONENTS,
-                TargetedBuildAction,
-                resolve_build_ref,
-            )
-
-            component = str(req.get("component") or "").strip().lower()
-            if component not in _COMPONENTS:
-                # A from-source framework build is the safest default for an
-                # arch/model request that named no valid compiled component.
-                component = "vllm_source"
-            framework = (getattr(state, "framework", "") or "").strip().lower() or "vllm"
-            gpu_type = (getattr(state, "gpu_type", "") or "").strip().lower()
-            repo_url = str(req.get("repo_url") or "").strip()
-            ref = str(req.get("ref") or "").strip()
-            source_pr_url = ""
-            # A PR-like ``ref``/``repo_url`` carries its own repo + provenance.
-            candidate = ref or repo_url
-            if candidate:
-                _repo, _ref, _pr = resolve_build_ref(candidate, repo_url)
-                repo_url = _repo or repo_url
-                # Take the resolved ref verbatim, including empty. An empty ref
-                # means "not checkoutable, autoselect a tag" (an issue citation),
-                # so falling back to the raw request here would hand the builder
-                # back the very string resolution just rejected.
-                ref = _ref
-                source_pr_url = _pr
-            if not _repo_matches_targeted_build_component(repo_url, component):
-                log.warning(
-                    "ENABLEMENT: specialist-requested build repo/component mismatch component=%s repo_url=%s",
-                    component,
-                    repo_url,
-                )
-                _consume_marker()
-                return
-            capability = str(req.get("capability") or "").strip()
-            reason = str(req.get("reason") or "").strip() or "specialist-requested targeted build"
-            action = TargetedBuildAction(
-                gap_id=f"gap.enablement.{capability or component}",
-                framework=framework,
-                component=component,
-                capability=capability or component,
-                reason=f"specialist request: {reason}",
-                repo_url=repo_url,
-                ref=ref,
-                gpu_arch=_derive_gpu_arch(gpu_type),
-                build_budget_sec=0,
-                source_pr_url=source_pr_url,
-            )
-            build_task_id = await self.enqueue_targeted_build(action)
-            if build_task_id:
-                _consume_marker()
-                log.info(
-                    "ENABLEMENT: enqueued specialist-requested targeted_build "
-                    "component=%s capability=%s ref=%s task=%s",
-                    component,
-                    capability or "(none)",
-                    ref or "(autoselect)",
-                    build_task_id,
-                )
-        except Exception:  # noqa: BLE001 — best-effort; never wedge dispatch
-            log.debug("enablement: specialist-requested build enqueue failed", exc_info=True)
-
-    async def _maybe_route_build_outcomes(self) -> None:
-        """Route terminal targeted_build rows to _maybe_rearm_enablement.
-
-        Called every tick from _pump_enablement_safely.  Reads succeeded/failed
-        rows, synthesises the rearm res dict (status='kept'/'reverted'/'advanced'),
-        and delegates to the existing stall-gate machinery.
-
-        Novelty: a 'timeout' or 'preflight_budget' failure_class maps to
-        'advanced' (novel attempt, time vs defect distinction — keep going); all
-        real defects map to 'reverted' (advance stall streak).
-
-        A succeeded build no longer synthesises status='kept' directly. Instead
-        it enqueues an integrate_patch launch probe so the runtime must actually
-        boot the model before KEEP is declared. Each row is read once, except that
-        a build whose probe was cancelled before it ran is read again: nothing
-        launched the runtime, so nothing has decided anything about that build.
-
-        Oldest-unrouted-first avoids starving older builds when a newer row is
-        already routed; the tradeoff is that a succeeded build whose probe cannot
-        be enqueued yet is retried every tick and can defer newer failed builds
-        until its probe opens or the budget recovers.
-        """
-        try:
-            all_tasks = []
-            for st in ("succeeded", "failed"):
-                all_tasks.extend(t for t in await self.tasks.by_state(st) if t.kind == "targeted_build")
-            if not all_tasks:
-                return
-            # Oldest terminal row first so a newer, already-routed build cannot
-            # hide an older build that still needs routing.
-            for task in sorted(
-                all_tasks,
-                key=lambda t: str(getattr(t, "updated_at", "") or ""),
-            ):
-                task_id = str(getattr(task, "task_id", "") or "")
-                # Skip rows already accounted for (tracked by enablement_build_manifest),
-                # unless what they were routed to was cancelled before it ran, which
-                # leaves the build no more launched than an unrouted one.
-                routed = self._build_routing_record(task_id)
-                if routed is not None and not await self._build_probe_was_cancelled(routed):
-                    continue
-
-                if task.state == "succeeded":
-                    await self._route_succeeded_build(task, routed)
-                else:
-                    await self._route_failed_build(task)
-                return
-        except Exception:  # noqa: BLE001 — never wedge the tick
-            log.debug("enablement: route_build_outcomes failed", exc_info=True)
-
-    async def _route_failed_build(self, task: "Task") -> None:
-        """Route a failed targeted_build row through the enablement rearm path."""
-        task_id = str(getattr(task, "task_id", "") or "")
-        state = self.shared_state
-        fc = ""
-        # failed row — read failure_class from history or last_build_failure
-        history = getattr(task, "history", None) or []
-        if isinstance(history, (list, tuple)) and history:
-            last_ev = history[-1]
-            if isinstance(last_ev, dict):
-                fc = str(last_ev.get("evidence", {}).get("failure_class") or "")
-
-        lbf = state.enablement.last_build_failure or {}
-        if not fc and isinstance(lbf, dict):
-            fc = str(lbf.get("failure_class") or "")
-
-        # Novelty ledger: time-based failures are always advanced; defect
-        # failures are advanced when the (component,ref,gpu_arch,cmd) tuple
-        # has not been seen before (novel), reverted when it is a repeat.
-        time_classes = frozenset({"timeout", "preflight_budget", "preflight_disk", "preflight_toolchain"})
-        novelty_key: list[Any] | None = None
-        if fc in time_classes:
-            new_log = str(state.enablement.launch_log or "")
-            res = {
-                "enablement": True,
-                "status": "advanced",
-                "advanced": True,
-                "patches_applied": [],
-                "enablement_launch_log": new_log,
-            }
-        else:
-            from ..framework.build_actions import TargetedBuildAction as _TBA, build_novelty_key as _bnk
-
-            task_params = getattr(task, "params", None) or {}
-            _action = _TBA.from_state(task_params)
-            _key = list(_bnk(_action))
-            ledger = list(state.enablement.build_novelty or [])
-            is_repeat = any(entry == _key for entry in ledger if isinstance(entry, list))
-            if is_repeat:
-                res = {"enablement": True, "status": "reverted"}
-                novelty_key = None
-            else:
-                new_log = str(state.enablement.launch_log or "")
-                res = {
-                    "enablement": True,
-                    "status": "advanced",
-                    "advanced": True,
-                    "patches_applied": [],
-                    "enablement_launch_log": new_log,
-                }
-                novelty_key = _key
-        log.info(
-            "ENABLEMENT: targeted_build %s task=%s failure_class=%r",
-            res["status"],
-            task_id,
-            fc,
-        )
-        # Rearm, ledger append, and manifest ack must stay together: a failed
-        # rearm leaves the build unrouted and the novelty ledger unchanged.
-        self._maybe_rearm_enablement(res)
-        if novelty_key is not None:
-            ledger = list(state.enablement.build_novelty or [])
-            ledger.append(novelty_key)
-            state.enablement.build_novelty = ledger[-20:]
-        self._note_build_routed(task_id)
-
-    async def _route_succeeded_build(self, task: "Task", routed: dict[str, Any] | None) -> None:
-        """Turn a succeeded targeted build into a launch probe, or a no-progress round.
-
-        KEEP is declared by the probe, never here: an artifact that builds is not
-        a runtime that boots. So this either opens a probe -- and remembers which
-        one, because a probe cancelled before it ran leaves the build unlaunched
-        and worth another -- or, when the built runtime cannot even be read, ends
-        the round as a revert.
-
-        The build is recorded as accounted for only once there is something to
-        account for. A probe the budget refused leaves it unrouted on purpose: it
-        is still a probed build nothing has launched, and the manifest saying
-        otherwise is what would strand it for the rest of the session.
-
-        Args:
-            task: The succeeded ``targeted_build`` row.
-            routed: What this build was routed to before, when it is being routed
-                again after its probe was cancelled; ``None`` on the first pass.
-        """
-        task_id = str(getattr(task, "task_id", "") or "")
-        attempt_root = str((getattr(task, "params", {}) or {}).get("attempt_root") or "")
-        # The build's attempt_root is resolved at pump time and is NOT written
-        # back into the task params (they keep the enqueue-time default ""). Fall
-        # back to the deterministic build path keyed by task_id so a *successful*
-        # build is not wrongly rejected as ``artifact_unreadable`` (mirrors
-        # BuildLifecycle._attempt_root).
-        if not attempt_root and task_id:
-            attempt_root = str(self.session_dir / "enablement" / "builds" / task_id)
-        br = None
-        if attempt_root:
-            from ..framework.targeted_build import _load_result_json
-
-            br = _load_result_json(attempt_root)
-
-        # If the runtime can't be read, it can't be launched → reverted.
-        if br is None or not br.ok or not br.runtime.to_runtime_override():
-            log.info("ENABLEMENT: targeted_build artifact-unreadable task=%s", task_id)
-            self._maybe_rearm_enablement({"enablement": True, "status": "reverted", "reason": "artifact_unreadable"})
-            self._note_build_routed(task_id)
-            return
-
-        log.info("ENABLEMENT: targeted_build probe-verified → enqueue launch probe task=%s", task_id)
-        probe_tid, generation = await self._enqueue_build_launch_probe(
-            task_id,
-            br,
-            generation=int((routed or {}).get("probe_generation") or 0),
-        )
-        if not probe_tid:
-            return
-        self._note_build_routed(task_id, probe_task_id=probe_tid, probe_generation=generation)
-
-    def _build_routing_record(self, build_task_id: str) -> dict[str, Any] | None:
-        """The record of what a build's outcome was already routed to, if any.
-
-        Only the routing sentinels carry a ``task_id``; the build attempts the
-        lifecycle appends to the same manifest carry ``ok`` instead.
-
-        Args:
-            build_task_id: The ``targeted_build`` row to look for.
-
-        Returns:
-            The sentinel dict, or ``None`` when this build has not been routed.
-        """
-        for entry in reversed(list(self.shared_state.enablement.build_manifest or [])):
-            if isinstance(entry, dict) and str(entry.get("task_id") or "") == build_task_id:
-                return entry
-        return None
-
-    def _note_build_routed(self, build_task_id: str, **fields: Any) -> None:
-        """Record that a build's outcome has been routed, and to what.
-
-        Args:
-            build_task_id: The ``targeted_build`` row that was routed.
-            fields: What it was routed to, for a build whose outcome is a probe.
-        """
-        manifest = list(self.shared_state.enablement.build_manifest or [])
-        for idx, entry in enumerate(manifest):
-            if isinstance(entry, dict) and str(entry.get("task_id") or "") == build_task_id:
-                manifest[idx] = {**entry, **fields}
-                break
-        else:
-            manifest.append({"task_id": build_task_id, "routed": True, **fields})
-        self.shared_state.enablement.build_manifest = manifest
-
-    async def _build_probe_was_cancelled(self, routed: dict[str, Any]) -> bool:
-        """Whether the probe a build was routed to was stopped before it ran.
-
-        A cancelled probe is no evidence about the build: the queue scan drops a
-        queued row the wall-clock budget can no longer fit, and a phase boundary
-        drops one the new phase does not allow. Either way the built runtime was
-        never launched, so the build is still owed a probe -- and without noticing
-        that, the manifest entry written when the first one was opened keeps this
-        build accounted for permanently, across resumes included.
-
-        Args:
-            routed: The build's routing record.
-
-        Returns:
-            ``True`` when the recorded probe exists and was cancelled.
-        """
-        from ..state.task_registry import TaskNotFound
-
-        probe_tid = str(routed.get("probe_task_id") or "").strip()
-        if not probe_tid:
-            return False
-        try:
-            probe = await self.tasks.get(probe_tid)
-        except TaskNotFound:
-            # Pruned rather than cancelled; re-probing on a row that is gone
-            # would re-probe on every later tick too.
-            return False
-        return str(getattr(probe, "state", "") or "") == "cancelled"
-
-    async def _enqueue_build_launch_probe(
-        self,
-        build_task_id: str,
-        br: Any,
-        *,
-        generation: int = 0,
-    ) -> tuple[str, int]:
-        """Enqueue an integrate_patch launch probe for a verified build.
-
-        Runs the built runtime through the enablement runnable gate without
-        applying any patch.  The probe completes as an ordinary integrate_patch
-        task whose enablement:True result is routed by the dispatcher through
-        _maybe_rearm_authored_lane → _maybe_rearm_enablement, producing a
-        genuine KEEP/advanced/reverted outcome.  The whole-machine GPU pool is
-        acquired via _framework_gpu_params.
-
-        The probe is what declares KEEP for a build, so it must not be opened
-        into a session that cannot run it: the queue scan drops a queued row the
-        wall-clock budget can no longer fit, and a probe cancelled that way
-        leaves the build verified but never launched. So the same gate the scan
-        asks is asked here first, and a denial enqueues nothing -- the build
-        stays unrouted, and the tick or resume that can afford a probe opens one.
-
-        Args:
-            build_task_id: The verified build this probe launches.
-            br: Its ``BuildResult``, read for the runtime override.
-            generation: The probe generation to try first, from what this build
-                was routed to before.
-
-        Returns:
-            The probe ``task_id`` and the generation it sits on; the id is empty
-            when nothing was enqueued.
-        """
-        from hyperloom.agents.framework.enablement import classify_failure
-
-        denied = self._time_budget_denial_for_action("integrate_patch")
-        if denied is not None:
-            log.info(
-                "ENABLEMENT: build launch probe held for build=%s, not enqueued -- %s",
-                build_task_id,
-                denied,
-            )
-            return "", generation
-        state = self.shared_state
-        runtime_override = br.runtime.to_runtime_override()
-        launch_log = str(state.enablement.launch_log or "")
-        before_sig = classify_failure(launch_log).to_dict()
-        params: dict[str, Any] = {
-            "enablement": True,
-            "enablement_launch_only": True,
-            "runtime_override": runtime_override,
-            "framework": str(getattr(state, "framework", "") or ""),
-            "enablement_before_signature": before_sig,
-            "source": "coordinator_internal",
-            **self._framework_gpu_params(),
-            **_enablement_carrier_params(state),
-        }
-        # Prefer the eval-origin probe config so the re-run keeps the original
-        # workload/eval contract; fall back to the promoted baseline config.
-        cfg = str(state.enablement.probe_config_path or "") or str(getattr(state, "baseline_config_path", "") or "")
-        if cfg:
-            params["config_path"] = cfg
-        lanes, ttl = self._framework_authoring_lanes_ttl(params, base_ttl_sec=3600)
-        probe_task, generation = await self._open_row_past_spent_generations(
-            kind="integrate_patch",
-            params=params,
-            key_for=lambda gen: f"build_launch_probe:{build_task_id}:gen{gen}",
-            generation=generation,
-            label="build launch probe",
-            requires_lanes=lanes,
-            lease_ttl_sec=ttl,
-        )
-        if probe_task is None:
-            return "", generation
-        probe_tid = str(getattr(probe_task, "task_id", "") or "")
-        log.info(
-            "ENABLEMENT: build launch probe task=%s gen%d state=%s (build=%s)",
-            probe_tid,
-            generation,
-            getattr(probe_task, "state", ""),
-            build_task_id,
-        )
-        return probe_tid, generation
-
-    async def _maybe_enqueue_enablement_baseline_revalidation(self) -> str:
-        """Enqueue one genuine baseline to revalidate a KEEP'd eval-origin patch.
-
-        Uses the accepted config from the KEEP'd candidate bench (preferred) or
-        falls back to the original probe config, plus that bench's env/arg layers,
-        which the YAML does not carry and without which a different configuration
-        would be graded. The frozen eval controls from the carrier params ensure
-        RUN_EVAL and eval task/limit match the trigger contract. Idempotent and
-        one-at-a-time.
-        """
-        state = self.shared_state
-        if not bool(state.enablement.validation_pending):
-            return ""
-        # If we already have a tracked revalidation task that is still alive, do
-        # not create another one.
-        tracked_tid = str(state.enablement.revalidation_task_id or "").strip()
-        if tracked_tid:
-            try:
-                for t in (*await self.tasks.queued(), *await self.tasks.running()):
-                    if str(getattr(t, "task_id", "") or "") == tracked_tid:
-                        return tracked_tid
-            except Exception:  # noqa: BLE001 — defensive
-                pass
-        # Do not open a row the dispatcher would cancel on sight. A revalidation
-        # is a full baseline, and the queue scan drops a queued one the session
-        # budget can no longer fit -- which leaves a cancelled row owning this
-        # window's idempotency key, and a row cancelled at dispatch never
-        # produces a result to route, so nothing would advance the generation
-        # past it. Holding the window shut for now costs nothing: it stays open,
-        # and the resume that has budget again enqueues it.
-        denied = self._time_budget_denial_for_action("baseline")
-        if denied is not None:
-            log.info("ENABLEMENT revalidation: window held open, not enqueued -- %s", denied)
-            return ""
-        params: dict[str, Any] = {
-            "source": "coordinator_internal",
-            "reason": ENABLEMENT_REVALIDATION_REASON,
-            "disable_run_eval": False,
-            **_enablement_carrier_params(state),
-        }
-        accepted_cfg = str(state.enablement.accepted_config_path or "").strip()
-        probe_cfg = str(state.enablement.probe_config_path or "").strip()
-        cfg = accepted_cfg or probe_cfg
-        if cfg:
-            params["config_path"] = cfg
-        effective = state.enablement.accepted_config
-        for key in ("extra_envs", "extra_server_args", "remove_args", "unset_envs", "args_mode"):
-            if effective.get(key):
-                params[key] = effective[key]
-        # Carry the active runtime override so the revalidation baseline runs
-        # under the same framework runtime as the KEEP'd candidate.
-        active_rt = state.enablement.active_runtime or {}
-        if isinstance(active_rt, dict) and active_rt:
-            from ..framework.stack_actions import FrameworkRuntime
-
-            rt_obj = FrameworkRuntime.from_state(active_rt)
-            rt_override = rt_obj.to_runtime_override()
-            if rt_override:
-                params["runtime_override"] = rt_override
-        task = await self._open_revalidation_row(params)
-        if task is None:
-            return ""
-        task_id = str(getattr(task, "task_id", "") or "")
-        # Persist the task_id so _promote_baseline can verify identity.
-        if task_id and task_id != str(state.enablement.revalidation_task_id or ""):
-            state.enablement.revalidation_task_id = task_id
-            try:
-                state.save(self.session_dir)
-            except Exception:  # noqa: BLE001 — defensive
-                log.debug("enablement revalidation: save of task_id failed", exc_info=True)
-        return task_id
-
-    async def _open_revalidation_row(self, params: dict[str, Any]) -> "Task | None":
-        """Resolve this revalidation window's task row, on a generation it can use.
-
-        Args:
-            params: The baseline params for the revalidation row.
-
-        Returns:
-            The row to track, or ``None`` when this tick found only spent
-            generations -- the window stays open and the next tick tries again.
-        """
-        state = self.shared_state
-        _baseline_lanes, _baseline_ttl = self._registry_lanes_ttl("baseline")
-        task, generation = await self._open_row_past_spent_generations(
-            kind="baseline",
-            params=params,
-            key_for=lambda gen: f"enablement_revalidation:gen{gen}",
-            generation=int(state.enablement.revalidation_generation or 0),
-            label="revalidation",
-            # Both halves of the catalogue contract: a baseline re-launches the
-            # server, so it must hold the same lanes any other baseline does.
-            requires_lanes=_baseline_lanes,
-            lease_ttl_sec=_baseline_ttl,
-        )
-        state.enablement.revalidation_generation = generation
-        return task
-
-    async def _open_row_past_spent_generations(
-        self,
-        *,
-        kind: str,
-        params: dict[str, Any],
-        key_for: Callable[[int], str],
-        generation: int,
-        label: str,
-        attempts: int = 2,
-        **create_kwargs: Any,
-    ) -> tuple["Task | None", int]:
-        """Create or re-use a task row, skipping generations already spent.
-
-        A generation in the idempotency key is what lets one piece of work get a
-        fresh row after an earlier attempt at it went terminal. That only holds if
-        a key resolving to a terminal row is recognised as a spent generation
-        rather than an enqueue: a row cancelled at dispatch -- which is what the
-        queue scan does to work the wall-clock budget can no longer fit -- never
-        produces a result to route, so nothing downstream advances the generation
-        past it, and every later attempt resolves to a row that measured nothing.
-
-        Args:
-            kind: The task kind to create.
-            params: The task params.
-            key_for: Builds the idempotency key for a generation number.
-            generation: The generation to try first.
-            label: How this work is named in the log when a generation is spent.
-            attempts: How many generations to try before giving up this pass.
-            **create_kwargs: Passed through to
-                :meth:`TaskRegistry.create_or_return_existing` (lanes, TTL).
-
-        Returns:
-            The row to use and the generation it sits on, or ``None`` with the
-            generation to try next when every attempt this pass found a spent
-            one. The caller persists the generation, since only the caller knows
-            where it lives.
-        """
-        from ..state.task_registry import TERMINAL_STATES
-
-        for _attempt in range(max(1, attempts)):
-            task, _existing = await self.tasks.create_or_return_existing(
-                kind=kind,
-                params=params,
-                idempotency_key=key_for(generation),
-                **create_kwargs,
-            )
-            if str(getattr(task, "state", "") or "") not in TERMINAL_STATES:
-                return task, generation
-            log.warning(
-                "ENABLEMENT %s: gen%d resolves to terminal task %s (%s); opening "
-                "generation %d so the work is not stuck on it",
-                label,
-                generation,
-                getattr(task, "task_id", ""),
-                getattr(task, "state", ""),
-                generation + 1,
-            )
-            generation += 1
-        return None, generation
-
-    async def _pump_enablement_safely(self, *, caller: str) -> None:
-        """Phase-independent enablement pump — runs every tick.
-
-        A baseline that cannot even *launch* traps the run in PRELUDE forever:
-        the only PRELUDE exit gate is ``baseline_tput > 0``, which a
-        non-runnable (model, backend) combo never reaches. The enablement
-        authoring dispatch used to live only inside
-        :meth:`_pump_framework_agent_phase` (guarded on
-        ``phase == FRAMEWORK_AGENT``), so it could never fire for the exact
-        "can't boot at all" scenario it exists to repair — the run instead hit
-        the 3-failure ``baseline_failed`` stop.
-
-        This wrapper drives :meth:`_maybe_enqueue_enablement_specialist` from
-        every coordinator tick, independent of phase. All dispatch guards
-        (dispatched-in-flight, already-succeeded, ``baseline_tput > 0``,
-        failure-streak, run deadline, single-node) live inside that method, so
-        calling it unconditionally here is safe and idempotent.
-
-        Args:
-            caller: Label identifying the caller ("tick" / "run"), for logs.
-        """
-        try:
-            await self._maybe_route_build_outcomes()
-        except Exception:  # noqa: BLE001 — never wedge the tick
-            log.exception("ENABLEMENT route_build_outcomes (%s) failed", caller)
-        try:
-            await self._maybe_enqueue_enablement_baseline_revalidation()
-        except Exception:  # noqa: BLE001 — never wedge the tick
-            log.exception("ENABLEMENT revalidation (%s) failed", caller)
-        try:
-            await self._maybe_enqueue_enablement_specialist()
-        except Exception:  # noqa: BLE001 — never wedge the tick
-            log.exception("ENABLEMENT pump (%s) failed", caller)
 
     def _record_framework_agent_authored_outcome(
         self,
